+++ src/com/coryginsberg/Managers/FacilityManager.java
@@ -17,7 +17,7 @@ public class FacilityManager {
 
     private static ArrayList<Facility> facilities = new ArrayList<>();
 
-    public void addFacility(String name, int rate, int cost, ArrayList<HashMap<Float, String>> linkedCities) {
+    public void addFacility(String name, int rate, int cost, ArrayList<HashMap<Integer, String>> linkedCities) {
         Facility facility = new Facility(name, rate, cost, linkedCities);
         facilities.add(facility);
     }
